Move PeerPage to folder
