+++ frontend/src/App.tsx
@@ -8,7 +8,7 @@ import { AnimatePresence } from "framer-motion";
 import Menu from "./components/Menu";
 import DashboardPage from "./components/Dashboard/DashboardPage";
 import FileUploadPage from "./components/FileUpload/FileUploadPage";
-import PeerPage from "./components/PeerPage";
+import PeerPage from "./components/PeerPage/PeerPage";
 import SettingPage from "./components/Setting/SettingPage";
 
 function LocationProvider({ children }: { children: React.ReactNode }) {
